refactor: Dto 분리
